+++ kensho_kenverters/convert_output.py
@@ -30,10 +30,11 @@ def table_to_markdown(table: TableType) -> str:
         row = [str(x) for x in row]
         line = "| " + " | ".join(row) + " |"
         lines.append(line)
-        # Markdown tables only render if there's a header row
+        # Markdown tables only render correctly if there's a header row
+        # and an extra line at the end to mark the end of the table
         if row_index == 0:
             lines.append("| " + " | ".join(["---" for _ in row]) + " |")
-    table_text = "\n".join(lines)
+    table_text = "\n".join(lines) + "\n"
     return table_text